clean up favicons, config improvements
