+++ packages/atlas/src/index.html
@@ -10,11 +10,7 @@
     <link rel="icon" href="/favicon.ico" sizes="any" />
     <link rel="icon" href="/icon.svg" type="image/svg+xml" />
     <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png" />
-    <link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png" />
-    <link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png" />
     <link rel="manifest" href="/manifest.webmanifest" />
-    <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#4038ff" />
-    <meta name="msapplication-TileColor" content="#4038ff" />
     <meta name="theme-color" content="#000" />
 
     <!--  fonts imports  -->
